Update the Human Pet
This Pet Type has currently not been added to the game, as I have not
yet discovered a way to successfully send the reuquired information to
the clients about skins. Even the Steve skin doesn't work yet. It
throws errors, as the "Monster" entity (ID 49) no longer has a texture
client-side.