Pets now fly when their owner does
Pets generally try to stay behind the Player's field of vision to
prevent obstruction. As a result, it is near impossible to view (and
consequently open the Pet Menu) while flying, as the Pet will always
avoid your line of sight.

They're a little bit shy of their new flying ability <3