Improve SQL support.
This is an essential rewrite of the SQL support, which I believe to be better than the existing support.

Connection pooling is now implemented, thanks to BoneCP. It's shaded in along with SLF4J, but I'm open to other suggestions.

Additionally, the SQL serialization support has been rewritten to enhance security and allow the use of the ' character.