Experimental: Sponge and Bukkit compatibility
This is essentially an entirely new set of modules. To put it simply,
the API is now bridged between two platforms by a ‘Bridge’ module,
which is then utilised each of the plugin modules. Everything that is
not platform-specific is housed in either the Bridge or API module.

A few new libraries are also made use of extensively (Influx [with
Sponge support], Commodus v2, Interact).

WIP/EXPERIMENTAL: DOES NOT COMPILE